backend/requirements.txt
# backend/requirements.txt

# ── Web framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # file uploads

# ── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0                # async PostgreSQL driver
psycopg2-binary==2.9.9         # sync driver (Alembic, Celery)
alembic==1.13.1
pgvector==0.2.5                # pgvector SQLAlchemy type

# ── Auth ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0                  # for GitHub OAuth exchange

# ── Task queue ─────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4
flower==2.0.1                  # Celery monitoring

# ── Config ─────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── CV parsing ─────────────────────────────────────────────
pdfplumber==0.11.1
PyMuPDF==1.24.3
python-docx==1.1.2
spacy==3.7.4
# After install, run: python -m spacy download en_core_web_lg
rapidfuzz==3.9.0
dateparser==1.2.0

# ── GitHub ─────────────────────────────────────────────────
PyGithub==2.3.0

# ── ML & Embeddings ────────────────────────────────────────
sentence-transformers==2.7.0
scikit-learn==1.4.2
xgboost==2.0.3
numpy==1.26.4
torch==2.3.0                   # CPU-only for Railway; use GPU variant for AWS

# ── LLM / Roadmap ──────────────────────────────────────────
openai==1.30.1
anthropic==0.26.0
langchain==0.2.1
langchain-openai==0.1.7

# ── Object storage ─────────────────────────────────────────
boto3==1.34.115

# ── Email ──────────────────────────────────────────────────
sendgrid==6.11.0

# ── Monitoring ─────────────────────────────────────────────
sentry-sdk[fastapi]==2.3.1
prometheus-fastapi-instrumentator==6.1.0

# ── Testing ────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0                  # TestClient
factory-boy==3.3.0